requirements.txt
numpy>=1.24.0,<2.0.0
tensorflow>=2.13.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-learn>=1.3.0

# GPU support (optional but recommended)
# For NVIDIA GPUs with CUDA support:
# tensorflow[and-cuda]>=2.13.0
#
# Note: GPU training requires:
# - NVIDIA GPU with CUDA support
# - CUDA Toolkit 11.8 or compatible
# - cuDNN 8.6 or compatible
#
# Installation for GPU:
# pip install tensorflow[and-cuda]>=2.13.0

